backend/requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
mysql-connector-python
python-dotenv
werkzeug
pandas
openpyxl
requests

